requirements.txt
fastapi
uvicorn
python-jose
databases[postgresql]
sqlalchemy
alembic
psycopg2-binary
passlib
python-multipart
python-dotenv
PyJWT
pydantic[email]
bcrypt
pytesseract 
pdf2image 
langchain 
langchain_community 
langchain-openai 
ollama
pypdf
pytest 
httpx
pytest_asyncio